+++ trans/context_processors.py
@@ -23,4 +23,5 @@ def url(request):
 def mt(request):
     return {
         'apertium_api_key': settings.MT_APERTIUM_KEY,
+        'microsoft_api_key': settings.MT_MICROSOFT_KEY,
     }