+++ docs/changes.rst
@@ -6,7 +6,7 @@ weblate 0.5
 
 Relased on ? 2012.
 
-* Support for machine translation using Apertium.
+* Support for machine translation using Apertium and Microsoft Translator.
 * Several new translations.
 * Improved merging of upstream changes.
 * Better handle concurrent git pull and translation.